# spanconfig,kv: merge adjacent ranges with identical configs 