--- a/pkg/server/server.go
+++ b/pkg/server/server.go
@@ -565,6 +565,7 @@ func NewServer(cfg Config, stopper *stop.Stopper) (*Server, error) {
 				keys.SpanConfigurationsTableID,
 				1<<20, /* 1 MB */
 				fallbackConf,
+				cfg.Settings,
 				spanConfigKnobs,
 			)
 		}