--- a/pkg/server/server_sql.go
+++ b/pkg/server/server_sql.go
@@ -990,6 +990,7 @@ func newSQLServer(ctx context.Context, cfg sqlServerArgs) (*SQLServer, error) {
 			execCfg,
 			codec,
 			cfg.TenantID,
+			cfg.Settings,
 			spanConfigKnobs,
 		)
 		spanConfig.manager = spanconfigmanager.New(