--- a/pkg/spanconfig/spanconfigstore/BUILD.bazel
+++ b/pkg/spanconfig/spanconfigstore/BUILD.bazel
@@ -1,3 +1,4 @@
+load("//pkg/util/interval/generic:gen.bzl", "gen_interval_btree")
 load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
@@ -6,6 +7,7 @@ go_library(
         "span_store.go",
         "store.go",
         "system_store.go",
+        ":spanconfigstoreentry_interval_btree.go",  # keep
     ],
     importpath = "github.com/cockroachdb/cockroach/pkg/spanconfig/spanconfigstore",
     visibility = ["//visibility:public"],
@@ -15,10 +17,10 @@ go_library(
         "//pkg/settings",
         "//pkg/settings/cluster",
         "//pkg/spanconfig",
-        "//pkg/util/interval",
         "//pkg/util/iterutil",
         "//pkg/util/log",
         "//pkg/util/syncutil",
+        "//pkg/util/timeutil",  # keep
         "@com_github_cockroachdb_errors//:errors",
     ],
 )
@@ -29,6 +31,7 @@ go_test(
         "span_store_test.go",
         "store_test.go",
         "system_store_test.go",
+        ":spanconfigstoreentry_interval_btree_test.go",  # keep
     ],
     data = glob(["testdata/**"]),
     embed = [":spanconfigstore"],
@@ -46,4 +49,10 @@ go_test(
         "@com_github_cockroachdb_datadriven//:datadriven",
         "@com_github_stretchr_testify//require",
     ],
+)
+
+gen_interval_btree(
+    name = "spanconfig_interval_btree",
+    package = "spanconfigstore",
+    type = "*spanConfigStoreEntry",
 )