--- a/pkg/spanconfig/spanconfigstore/BUILD.bazel
+++ b/pkg/spanconfig/spanconfigstore/BUILD.bazel
@@ -4,6 +4,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "spanconfigstore",
     srcs = [
+        "interner.go",
         "span_store.go",
         "store.go",
         "system_store.go",
@@ -19,6 +20,7 @@ go_library(
         "//pkg/spanconfig",
         "//pkg/util/iterutil",
         "//pkg/util/log",
+        "//pkg/util/protoutil",
         "//pkg/util/syncutil",
         "//pkg/util/timeutil",  # keep
         "@com_github_cockroachdb_errors//:errors",