… pkg/ccl/serverccl/statusccl/tenant_status_test.go
… pkg/ccl/serverccl/statusccl/tenant_test_utils.go
… pkg/kv/kvserver/client_spanconfigs_test.go
… pkg/spanconfig/spanconfigkvsubscriber/BUILD.bazel
… pkg/spanconfig/spanconfigkvsubscriber/datadriven_test.go
… pkg/spanconfig/spanconfigkvsubscriber/kvsubscriber_test.go
… pkg/spanconfig/spanconfigstore/entry_interval_btree.go
… pkg/spanconfig/spanconfigstore/entry_interval_btree_test.go
… pkg/spanconfig/spanconfigstore/span_store.go
… pkg/spanconfig/spanconfigstore/span_store_test.go
… pkg/spanconfig/spanconfigstore/spanconfigstoreentry_interval_btree.go
… pkg/spanconfig/spanconfigstore/spanconfigstoreentry_interval_btree_test.go
… pkg/spanconfig/spanconfigstore/store_test.go
… pkg/spanconfig/spanconfigstore/system_store_test.go
… pkg/spanconfig/spanconfigstore/testdata/batched/interned
… pkg/spanconfig/spanconfigstore/testdata/single/split_keys
… pkg/spanconfig/spanconfigtestutils/spanconfigtestcluster/cluster.go
… pkg/spanconfig/testing_knobs.go
… pkg/testutils/localtestcluster/local_test_cluster.go
… pkg/util/interval/interval.go
… pkg/util/interval/llrb_based_interval.go