@irfansharif 2022-04-13
This should be ready for review. There are some random out-of-package tests that need to be updated since they expect splits post-table creation (they were added after we introduced tenant range splits), and I also want to write a small benchmark for the read path. Like discussed offline, plan for this PR is to let it bake on master for a few days before backporting to 22.1, aiming for the .1 release. I also think this PR makes #78299 redundant (LMK if you disagree) -- so I'm just going to close that as well.