@irfansharif 2022-04-14
> There are some random out-of-package tests that need to be updated since they expect splits post-table creation (they were added after we introduced tenant range splits), and I also want to write a small benchmark for the read path.

Done and done.