@irfansharif 2022-04-20
Yup.

> plan for this PR is to let it bake on master for a few days before backporting to 22.1, aiming for the .1 release.

For single-tenant clusters it won't be enabled by default; there'll be an opt-in cluster setting.