@ajwerner 2022-04-29
While you're at it, give the `string` a strong type like `spanConfigKey`