@theodore-hyman 2022-05-02
@irfansharif Thank you for your time and attention on this matter. I am chiming in just to confirm the timeline here as a quick sanity check. 

Again, I know you just said that this would be backported to v22.1.1 and you just said: 

`I'll open the backport now but not merge to the release until some point mid next week` 

However just re-confirming, apologies for the redundancy. This is related to a highly escalated customer situation as you're likely aware.

My understanding is that this fix, in the form of a new hidden config option, will be released as part of v22.1.1 that is scheduled for public release June 6, 2022. Please confirm or let me know if I am misunderstanding. Thank you