@irfansharif 2022-05-03
The backport (#80811) was merged, so it'll make it to 22.1.1 whenever that goes out. That said, I don't know which customer you're referring to (DM?) and whether this would help. This PR wasn't intended as a fix for anything, so let's discuss elsewhere what we're expecting this to help with.